# Make >< build its result eagerly